+++ euporie/convert/formats/markdown.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import logging
 from typing import TYPE_CHECKING
 
 from euporie.convert.base import register
@@ -9,6 +10,10 @@ from euporie.convert.util import have_modules
 
 if TYPE_CHECKING:
     from typing import Optional
+
+log = logging.getLogger(__name__)
+
+_HTML2TEXT_TABLE_FIRST_CELL_RE = r"^((\|\s+[^\|]+\s+)((\|\s+[^\|]+\s+|:?-+:?\|)(\|\s+[^\|]+\s+|:?-+:?\|))*:?-+:?\|:?-+:?\s*$)"  # noqa B950
 
 
 @register(
@@ -24,10 +29,18 @@ def html_to_markdown_py_html2text(
     bg: "Optional[str]" = None,
 ) -> "str":
     """Convert HTML to markdown tables using :py:mod:`html2text`."""
+    import re
+
     from html2text import HTML2Text  # type: ignore
 
     parser = HTML2Text()
-    return parser.handle(data)
+    result = parser.handle(data)
+
+    # Fix for html2text issue with empty first cells in tables
+    # https://github.com/Alir3z4/html2text/pull/380
+    result = re.sub(_HTML2TEXT_TABLE_FIRST_CELL_RE, r"|  \1", result, 0, re.MULTILINE)
+
+    return result
 
 
 @register(